+++ examples/fluids/qfunctions/advection2d.h
@@ -20,9 +20,7 @@
 #ifndef advection2d_h
 #define advection2d_h
 
-#ifndef __CUDACC__
-#  include <math.h>
-#endif
+#include <math.h>
 
 #ifndef M_PI
 #define M_PI    3.14159265358979323846