+++ examples/solids/qfunctions/manufactured-force.h
@@ -20,9 +20,7 @@
 #ifndef MANUFACTURED_H
 #define MANUFACTURED_H
 
-#ifndef __CUDACC__
-#  include <math.h>
-#endif
+#include <math.h>
 
 #ifndef PHYSICS_STRUCT
 #define PHYSICS_STRUCT