+++ interface/ceed-hip.c
@@ -14,8 +14,9 @@
 // software, applications, hardware, advanced system engineering and early
 // testbed platforms, in support of the nation's exascale computing imperative.
 
-#include <ceed-impl.h>
+#include <ceed/ceed.h>
 #include <ceed/hip.h>
+#include <ceed-impl.h>
 
 /**
   @brief Set HIP function pointer to evaluate action at quadrature points